Thrill Education - Mission Development systems
A set of scripts that I developed that together create a modular system that allows us to put together new games quickly and easily.

The game manager hosts a list of panels (or screens) that will automatically transition from one to the other once certain requirements have been met. I track these requirements using the SendEnum script which tells the GameManager when a requirement has been fulfilled.

The SpeechManager manages whenever dialogue appears in the game - complete with subtitles - and allows me to trigger events that are exclusive to that piece of dialogue via SpeechEvents.