DinoDeliveryService - Game Management
The GameManager navigational system for my game DinoDeliveryService that allows me to move a character around a grid-based world and a control system so that the player can decide where they want to go. (There is also some additional parts at the beginning that I added to reattach some assets when giving the UI a complete overhaul).